Re-implemented the submit_test method and functions
Use the approach in QE's fixtures. Directly pass a folder to the
`prepare_for_submission` method. I have to do it this way because
at the moment the dry_run function does not allow the created folder
with input file to be access directly.
Also added the get_castep_input_summary method to the `CastepCalculation`
class.