Updated base and relax workchains
The base workchain now takes the inputs of the calculation at the `calc`
input namespace. The relax workchain now also has a `calc` name space which
in practice will be passed to that of the bas workchain.
This way we seperate the workchain level inputs with the calculation level
inptus.