Allow CastepCalculation to received extra kpoints
In the old days, multiple variant of the `CastepCalculation` are
created for specific tasks. This was because CASTEP will internally
perform some 'mini-workflows' for the task and individual Calculation
works well. In the 1.0 days, the workflow will be controlled by
higher level `WorkChain`s and it make sense to made a single
calculation to interface with the code.

This commit moves the ability to take extra kpoints back to
`CastepCalculation`. So in theory, any CASTEP calculation will be able
to run with it (it was still possible to do that, but one needs to manually
create the kpoint inputs).

TODO: Think about what to do for transition state search and Pot1D.