Fix resorting forces are velocities
CASTEP will sort atomic species internally. In this plugin, the
output structures has resorted, but the same is not done for forces.
This means that the forces were previously in the wrong order for
structures not sorted to begin with (hence resorting is needed).
This commit fixes this issue. However, I still need to add a test
for it. Perhaps for NaCl.