Added option to bypass any checks for CastepRelaxWorkchain
Setting `pypass` to True will allow any checks to be skipped and
the calculation immeidated treated as "converged" in the context
of geometry optimisation. This can be used to run singlepoint
calculations under the same interface, or skip any explict
restart. Note that unfinished geometry optimisation may still be
restarted at `CastepBaseWorkChain` level.